requirements.txt
boto3==1.24.37
Brotli==1.0.9
python-snappy==0.6.1
tqdm==4.64.0
xxhash==3.0.0
zstd==1.5.2.5
